Update Book.js to have access to user
The borrow button in Book.js does not have access
to the currently logged in user

Let's update the functionality of the borrow button
in Book.js to have access to the currently logged
in user. This allows the addition of the borrowed
book to the user's borrowed books list.